+++ docs/make.jl
@@ -33,6 +33,7 @@ makedocs(
                 "lib/internals/builder.md",
                 "lib/internals/cross-references.md",
                 "lib/internals/docchecks.md",
+                "lib/internals/docmeta.md",
                 "lib/internals/docsystem.md",
                 "lib/internals/doctests.md",
                 "lib/internals/documenter.md",
@@ -40,6 +41,7 @@ makedocs(
                 "lib/internals/documents.md",
                 "lib/internals/dom.md",
                 "lib/internals/expanders.md",
+                "lib/internals/markdown2.md",
                 "lib/internals/mdflatten.md",
                 "lib/internals/selectors.md",
                 "lib/internals/textdiff.md",
@@ -50,6 +52,7 @@ makedocs(
         "contributing.md",
     ],
     strict = true,
+    doctest = ("doctest-only" in ARGS) ? :only : true,
 )
 
 deploydocs(