+++ src/DocChecks.jl
@@ -7,7 +7,8 @@ module DocChecks
 import ..Documenter:
     Documenter,
     Documents,
-    Utilities
+    Utilities,
+    Utilities.Markdown2
 
 using DocStringExtensions
 import Markdown
@@ -26,7 +27,7 @@ Prints out the name of each object that has not had its docs spliced into the do
 function missingdocs(doc::Documents.Document)
     doc.user.checkdocs === :none && return
     @debug "checking for missing docstrings."
-    bindings = allbindings(doc.user.checkdocs, doc.user.modules)
+    bindings = allbindings(doc.user.checkdocs, doc.blueprint.modules)
     for object in keys(doc.internal.objects)
         # The module references in docs blocks can yield a binding like
         # Docs.Binding(Mod, :SubMod) for a module SubMod, a submodule of Mod. However, the
@@ -107,7 +108,7 @@ function footnotes(doc::Documents.Document)
     # For all ids the final result should be `(N, 1)` where `N > 1`, i.e. one or more
     # footnote references and a single footnote body.
     footnotes = Dict{Documents.Page, Dict{String, Tuple{Int, Int}}}()
-    for (src, page) in doc.internal.pages
+    for (src, page) in doc.blueprint.pages
         empty!(page.globals.meta)
         orphans = Dict{String, Tuple{Int, Int}}()
         for element in page.elements
@@ -166,7 +167,7 @@ Checks external links using curl.
 function linkcheck(doc::Documents.Document)
     if doc.user.linkcheck
         if hascurl()
-            for (src, page) in doc.internal.pages
+            for (src, page) in doc.blueprint.pages
                 for element in page.elements
                     Documents.walk(page.globals.meta, page.mapping[element]) do block
                         linkcheck(block, doc)