+++ src/Writers/HTMLWriter.jl
@@ -223,7 +223,7 @@ end
 """
 Returns a page (as a [`Documents.Page`](@ref) object) using the [`HTMLContext`](@ref).
 """
-getpage(ctx, path) = ctx.doc.internal.pages[path]
+getpage(ctx, path) = ctx.doc.blueprint.pages[path]
 getpage(ctx, navnode::Documents.NavNode) = getpage(ctx, navnode.page)
 
 
@@ -250,7 +250,7 @@ function render(doc::Documents.Document, settings::HTML=HTML())
     push!(ctx.local_assets, copy_asset("documenter.css", doc))
     append!(ctx.local_assets, settings.assets)
 
-    for page in keys(doc.internal.pages)
+    for page in keys(doc.blueprint.pages)
         idx = findfirst(nn -> nn.page == page, doc.internal.navlist)
         nn = (idx === nothing) ? Documents.NavNode(page, nothing, nothing) : doc.internal.navlist[idx]
         @debug "Rendering $(page) [$(repr(idx))]"
@@ -520,7 +520,7 @@ function navitem(ctx, current, nn::Documents.NavNode)
 
     # add the subsections (2nd level headings) from the page
     if (nn === current) && current.page !== nothing
-        subs = collect_subsections(ctx.doc.internal.pages[current.page])
+        subs = collect_subsections(ctx.doc.blueprint.pages[current.page])
         internal_links = map(subs) do s
             istoplevel, anchor, text = s
             _li = istoplevel ? li[".toplevel"] : li[]
@@ -1162,7 +1162,7 @@ function fixlinks!(ctx, navnode, link::Markdown.Link)
     end
     path = normpath(joinpath(dirname(navnode.page), first(s)))
 
-    if endswith(path, ".md") && path in keys(ctx.doc.internal.pages)
+    if endswith(path, ".md") && path in keys(ctx.doc.blueprint.pages)
         # make sure that links to different valid pages are correct
         path = pretty_url(ctx, relhref(get_url(ctx, navnode), get_url(ctx, path)))
     elseif isfile(joinpath(ctx.doc.user.build, path))