+++ src/Writers/MarkdownWriter.jl
@@ -27,7 +27,7 @@ function render(doc::Documents.Document, settings::Markdown=Markdown())
     @info "MarkdownWriter: rendering Markdown pages."
     copy_assets(doc)
     mime = MIME"text/plain"()
-    for (src, page) in doc.internal.pages
+    for (src, page) in doc.blueprint.pages
         open(mdext(page.build), "w") do io
             for elem in page.elements
                 node = page.mapping[elem]