+++ test/doctests/doctests.jl
@@ -15,6 +15,7 @@ using Documenter.Utilities.TextDiff: Diff, Words
 
 include("src/FooWorking.jl")
 include("src/FooBroken.jl")
+include("src/NoMeta.jl")
 
 const builds_directory = joinpath(@__DIR__, "builds")
 ispath(builds_directory) && rm(builds_directory, recursive=true)
@@ -41,7 +42,15 @@ function run_makedocs(f, mdfiles, modules=Module[]; kwargs...)
     ------------------------------------ output ------------------------------------
     $(output)
     --------------------------------------------------------------------------------
-    """ result backtrace builddir
+    """ result stacktrace(backtrace) dir
+
+    write(joinpath(dir, "output"), output)
+    write(joinpath(dir, "output.onormalize"), onormalize(output))
+    open(joinpath(dir, "result"), "w") do io
+        show(io, "text/plain", result)
+        println(io, "-"^80)
+        show(io, "text/plain", stacktrace(backtrace))
+    end
 
     f(result, success, backtrace, output)
 end
@@ -62,6 +71,9 @@ function onormalize(s)
     # Remove filesystem paths in doctests failures
     s = replace(s, r"(doctest failure in )(.*)$"m => s"\1{PATH}")
     s = replace(s, r"(@ Documenter.DocTests )(.*)$"m => s"\1{PATH}")
+
+    # Remove stacktraces
+    s = replace(s, r"(│\s+Stacktrace:)(\n(│\s+)\[[0-9]+\].*)*" => s"\1\\n\3{STACKTRACE}")
 
     return s
 end
@@ -103,7 +115,7 @@ end
 
 rfile(filename) = joinpath(@__DIR__, "stdouts", filename)
 
-@testset "doctests" begin
+@testset "doctesting" begin
     # So, we have 4 doctests: 2 in a docstring, 2 in an .md file. One of either pair is
     # OK, other is broken. Here we first test all possible combinations of these doctest
     # with strict = true to make sure that the doctests are indeed failing.
@@ -136,7 +148,7 @@ rfile(filename) = joinpath(@__DIR__, "stdouts", filename)
 
     run_makedocs(["broken.md", "foobroken.md"]; modules=[FooBroken], strict=true) do result, success, backtrace, output
         @test !success
-        @test_broken is_same_as_file(output, rfile("stdout.6"))
+        @test is_same_as_file(output, rfile("stdout.6"))
     end
 
     run_makedocs(["fooworking.md"]; modules=[FooWorking], strict=true) do result, success, backtrace, output
@@ -145,8 +157,8 @@ rfile(filename) = joinpath(@__DIR__, "stdouts", filename)
     end
 
     run_makedocs(["foobroken.md"]; modules=[FooBroken], strict=true) do result, success, backtrace, output
-        @test_broken !success
-        @test_broken is_same_as_file(output, rfile("stdout.8"))
+        @test !success
+        @test is_same_as_file(output, rfile("stdout.8"))
     end
 
     # Here we try the default (strict = false) -- output should say that doctest failed, but
@@ -159,6 +171,45 @@ rfile(filename) = joinpath(@__DIR__, "stdouts", filename)
     run_makedocs(["broken.md"]) do result, success, backtrace, output
         @test success
         @test is_same_as_file(output, rfile("stdout.12"))
+    end
+
+    # Tests for doctest = :only. The outout should reflect that the docs themselves do not
+    # get built.
+    run_makedocs(["working.md"]; modules=[FooWorking], doctest = :only) do result, success, backtrace, output
+        @test success
+        @test is_same_as_file(output, rfile("stdout.21"))
+    end
+
+    run_makedocs(["working.md"]; modules=[FooBroken], doctest = :only) do result, success, backtrace, output
+        @test !success
+        @test is_same_as_file(output, rfile("stdout.22"))
+    end
+
+    run_makedocs(["broken.md"]; modules=[FooWorking], doctest = :only) do result, success, backtrace, output
+        @test !success
+        @test is_same_as_file(output, rfile("stdout.23"))
+    end
+
+    run_makedocs(["broken.md"]; modules=[FooBroken], doctest = :only) do result, success, backtrace, output
+        @test !success
+        @test is_same_as_file(output, rfile("stdout.24"))
+    end
+    # strict gets ignored with doctest = :only
+    run_makedocs(["broken.md"]; modules=[FooBroken], doctest = :only, strict=false) do result, success, backtrace, output
+        @test !success
+        @test is_same_as_file(output, rfile("stdout.25"))
+    end
+
+    # DocTestSetup in modules
+    run_makedocs([]; modules=[NoMeta], doctest = :only) do result, success, backtrace, output
+        @test !success
+        @test is_same_as_file(output, rfile("stdout.31"))
+    end
+    # Now, let's use Documenter's APIs to add the necessary meta information
+    DocMeta.setdocmeta!(NoMeta, :DocTestSetup, :(baz(x) = 2x))
+    run_makedocs([]; modules=[NoMeta], doctest = :only) do result, success, backtrace, output
+        @test success
+        @test is_same_as_file(output, rfile("stdout.32"))
     end
 end