+++ test/utilities.jl
@@ -222,7 +222,8 @@ end
     end
 
     import Documenter.Documents: Document, Page, Globals
-    let page = Page("source", "build", :build, [], IdDict{Any,Any}(), Globals()), doc = Document()
+    import Documenter.Utilities: Markdown2
+    let page = Page("source", "build", :build, [], IdDict{Any,Any}(), Globals(), Markdown2.MD()), doc = Document()
         code = """
         x += 3
         γγγ_γγγ